refactor: reuse feed page
