+++ packages/plugins/RSS3/src/SNSAdaptor/components/FeedCard/CommentCard.tsx
@@ -6,8 +6,9 @@ import type { FC } from 'react'
 import { Translate, useI18N } from '../../../locales/i18n_generated'
 import Markdown from 'react-markdown'
 import { useAddressLabel } from '../../hooks'
-import { CardFrame, CardType, FeedCardProps } from './base'
+import { CardFrame, FeedCardProps } from '../base'
 import { Label } from './common'
+import { CardType } from '../share'
 
 const useStyles = makeStyles<void, 'image' | 'verbose' | 'content'>()((theme, _, refs) => ({
     summary: {
