+++ packages/plugins/RSS3/src/SNSAdaptor/components/FeedCard/DonationCard.tsx
@@ -6,8 +6,9 @@ import { FC, HTMLProps, memo, useState } from 'react'
 import Markdown from 'react-markdown'
 import { Translate } from '../../../locales/i18n_generated'
 import { useAddressLabel } from '../../hooks'
+import { CardType } from '../share'
 import { Slider } from '../Slider'
-import { CardFrame, CardType, FeedCardProps } from './base'
+import { CardFrame, FeedCardProps } from '../base'
 import { formatValue, Label } from './common'
 
 const useStyles = makeStyles<void, 'image'>()((theme, _, refs) => ({
