+++ packages/plugins/RSS3/src/SNSAdaptor/components/FeedCard/TokenTransferCard.tsx
@@ -8,10 +8,10 @@ import { useI18N } from '../../../locales'
 import { Translate } from '../../../locales/i18n_generated'
 import { useFeedOwner } from '../../contexts'
 import { useAddressLabel } from '../../hooks'
-import { CardFrame, CardType, FeedCardProps } from './base'
+import { CardType } from '../share'
+import { CardFrame, FeedCardProps } from '../base'
 import { formatValue, Label } from './common'
 
-// const { Tag, Type } = RSS3BaseAPI
 const useStyles = makeStyles<void, 'tokenIcon' | 'verboseToken'>()((theme, _, refs) => ({
     summary: {
         fontSize: '14px',
