+++ packages/plugins/RSS3/src/SNSAdaptor/components/FeedCard/VoteCard.tsx
@@ -5,7 +5,8 @@ import type { FC } from 'react'
 import Markdown from 'react-markdown'
 import { Translate } from '../../../locales/i18n_generated'
 import { useAddressLabel } from '../../hooks'
-import { CardFrame, CardType, FeedCardProps } from './base'
+import { CardType } from '../share'
+import { CardFrame, FeedCardProps } from '../base'
 import { Label } from './common'
 
 const useStyles = makeStyles()((theme) => ({
