--- packages/plugins/RSS3/src/SNSAdaptor/components/ImageHolder.tsx
@@ -1,22 +1,0 @@
-import { Image } from '@masknet/shared'
-
-export interface ImageProps {
-    url: string
-    title?: string
-    isFullRound: boolean
-    size: number
-}
-
-export const ImageHolder = ({ url, title, isFullRound, size }: ImageProps) => {
-    const roundedClass = isFullRound ? 'rounded-full' : 'rounded'
-    const containerStyles = {
-        width: `${size}px`,
-        height: `${size}px`,
-    }
-
-    return (
-        <div className={`flex justify-around relative ${roundedClass}`} style={containerStyles}>
-            <Image className={`object-cover w-full h-full ${roundedClass}`} src={url} alt={title} />
-        </div>
-    )
-}
