+++ packages/plugins/RSS3/src/SNSAdaptor/components/base.tsx
@@ -1,12 +1,12 @@
 import { Icons } from '@masknet/icons'
-import type { GeneratedIconNonSquareProps } from '@masknet/icons/utils/internal'
 import { makeStyles } from '@masknet/theme'
 import type { RSS3BaseAPI } from '@masknet/web3-providers'
 import { Typography } from '@mui/material'
 import BigNumber from 'bignumber.js'
 import formatDateTime from 'date-fns/format'
-import type { ComponentType, FC, HTMLProps } from 'react'
-import { useViewFeedDetails } from '../../contexts/FeedDetails'
+import type { FC, HTMLProps } from 'react'
+import { useViewFeedDetails } from '../contexts'
+import { CardType, cardTypeIconMap, platformIconMap } from './share'
 
 export interface FeedCardBaseProps {
     feed: RSS3BaseAPI.Web3Feed
@@ -21,98 +21,6 @@ export interface FeedCardBaseProps {
 }
 
 export interface FeedCardProps extends Omit<HTMLProps<HTMLDivElement>, 'type' | 'action'>, FeedCardBaseProps {}
-
-export enum CardType {
-    AchievementBurn = 1,
-    AchievementReceive = 2,
-    CollectibleBurn = 3,
-    CollectibleIn = 4,
-    CollectibleMint = 5,
-    CollectibleOut = 6,
-    DonationDonate = 7,
-    DonationLaunch = 8,
-    GovernancePropose = 9,
-    GovernanceVote = 10,
-    NoteCreate = 11,
-    NoteEdit = 12,
-    NoteLink = 13,
-    NoteBurn = 14,
-    ProfileBurn = 15,
-    ProfileCreate = 16,
-    ProfileLink = 17,
-    TokenIn = 18,
-    TokenLiquidity = 19,
-    TokenOut = 20,
-    TokenStake = 21,
-    TokenSwap = 22,
-    UnknownBurn = 23,
-    UnknownCancel = 24,
-    UnknownIn = 25,
-    UnknownOut = 26,
-}
-
-type IconComponent = ComponentType<GeneratedIconNonSquareProps<never>>
-
-const iconMap: Record<CardType, IconComponent> = {
-    [CardType.AchievementBurn]: Icons.AchievementBurn,
-    [CardType.AchievementReceive]: Icons.AchievementReceive,
-    [CardType.CollectibleBurn]: Icons.CollectibleBurn,
-    [CardType.CollectibleIn]: Icons.CollectibleIn,
-    [CardType.CollectibleMint]: Icons.CollectibleMint,
-    [CardType.CollectibleOut]: Icons.CollectibleOut,
-    [CardType.DonationDonate]: Icons.DonationDonate,
-    [CardType.DonationLaunch]: Icons.DonationLaunch,
-    [CardType.GovernancePropose]: Icons.GovernancePropose,
-    [CardType.GovernanceVote]: Icons.GovernanceVote,
-    [CardType.NoteCreate]: Icons.NoteCreate,
-    [CardType.NoteEdit]: Icons.NoteEdit,
-    [CardType.NoteLink]: Icons.NoteLink,
-    [CardType.NoteBurn]: Icons.NoteBurn,
-    [CardType.ProfileBurn]: Icons.ProfileBurn,
-    [CardType.ProfileCreate]: Icons.ProfileCreate,
-    [CardType.ProfileLink]: Icons.ProfileLink,
-    [CardType.TokenIn]: Icons.TokenIn,
-    [CardType.TokenLiquidity]: Icons.TokenLiquidity,
-    [CardType.TokenOut]: Icons.TokenOut,
-    [CardType.TokenStake]: Icons.TokenStake,
-    [CardType.TokenSwap]: Icons.TokenSwap,
-    [CardType.UnknownBurn]: Icons.UnknownBurn,
-    [CardType.UnknownCancel]: Icons.UnknownCancel,
-    [CardType.UnknownIn]: Icons.UnknownIn,
-    [CardType.UnknownOut]: Icons.UnknownOut,
-}
-
-const platformIconMap: Record<RSS3BaseAPI.Network | RSS3BaseAPI.Platform, IconComponent | null> = {
-    // Networks
-    ethereum: Icons.ETH,
-    binance_smart_chain: Icons.ETH,
-    polygon: Icons.PolygonScan,
-    xdai: Icons.Gnosis,
-    arbitrum: Icons.Arbitrum,
-    optimism: Icons.Optimism,
-    fantom: Icons.Fantom,
-    avalanche: Icons.Avalanche,
-    // TODO icon for zksync is missing
-    zksync: Icons.ETH,
-    // Platforms
-    Gitcoin: Icons.Gitcoin,
-    Mirror: Icons.Mirror,
-    Snapshot: Icons.Snapshot,
-    Uniswap: Icons.Uniswap,
-    binance: Icons.BSC,
-    Lens: Icons.Mirror,
-    // TODO icon for zksync is missing
-    crossbell: Icons.Mirror,
-    '0x': Icons.ZeroX,
-    'ENS Registrar': null,
-    CrossSync: Icons.Lens,
-    Crossbell: Icons.Crossbell,
-    MetaMask: Icons.MetaMask,
-    OpenSea: Icons.OpenSea,
-    SushiSwap: null,
-    'crossbell.io': Icons.Crossbell,
-    xLog: Icons.XLog,
-}
 
 const useStyles = makeStyles()((theme) => ({
     inspectable: {
@@ -166,7 +74,7 @@ export const CardFrame: FC<CardFrameProps> = ({
     ...rest
 }) => {
     const { classes, cx } = useStyles()
-    const CardIcon = iconMap[type]
+    const CardIcon = cardTypeIconMap[type]
     const PrimaryPlatformIcon = feed.network ? platformIconMap[feed.network] : null
     const ProviderPlatformIcon = feed.platform ? platformIconMap[feed.platform] : null
