+++ packages/plugins/RSS3/src/SNSAdaptor/components/index.ts
@@ -1,6 +1,2 @@
-export * from './DonationCard.js'
-export * from './FootprintList.js'
-export * from './FootprintCard.js'
-export * from './ImageHolder.js'
 export * from './StatusBox.js'
 export * from './FeedCard/index.js'
