+++ packages/plugins/RSS3/src/SNSAdaptor/contexts/index.ts
@@ -1,1 +1,2 @@
 export * from './FeedOwnerContext.js'
+export * from './FeedDetails.js'
