--- packages/plugins/RSS3/src/SNSAdaptor/pages/index.ts
@@ -1,3 +1,0 @@
-export * from './DonationsPage.js'
-export * from './FeedsPage.js'
-export * from './FootprintsPage.js'
